Gui: Change SoQtOffscreenRendererPy to new PyCXX extension type
