+++ src/Gui/Application.cpp
@@ -84,7 +84,7 @@
 #include "SoFCDB.h"
 #include "Selection.h"
 #include "SelectionFilterPy.h"
-#include "SoFCOffscreenRenderer.h"
+#include "SoQtOffscreenRendererPy.h"
 #include "SplitView3DInventor.h"
 #include "TaskView/TaskView.h"
 #include "TaskView/TaskDialogPython.h"
